Split Handshake into submodules
Follow the same lines as with other mini-protocols.
