+++ ouroboros-network-framework/ouroboros-network-framework.cabal
@@ -32,6 +32,8 @@ library
 
                        Ouroboros.Network.Protocol.Handshake.Type
                        Ouroboros.Network.Protocol.Handshake.Codec
+                       Ouroboros.Network.Protocol.Handshake.Client
+                       Ouroboros.Network.Protocol.Handshake.Server
                        Ouroboros.Network.Protocol.Handshake.Version
                        Ouroboros.Network.Protocol.Handshake.Unversioned
                        Ouroboros.Network.Protocol.Limits
